Refactured javascript code to use javascript modules and other ES6 features
